erisx: convert to TS
